requirements.txt
python-dateutil
discord.py
APScheduler
pytz
requests

